+++ omapp4/src/main/java/br/com/tecsiscom/omapp/model/entity/estoque/ItemEstoque.java
@@ -32,7 +32,7 @@ public abstract class ItemEstoque implements Serializable{
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long id;
 	
-    private BigDecimal quantidade;
+    private Long quantidade;
 
 	@ManyToOne
 	@JoinColumn(nullable = false)
